Add GetPostById, parse ID and return JSON
Use int64 ID for posts and update storage interface. Implement
GetPostById: query and scan row, return ErrorNotFound for no rows. Parse
URL id in handler and write JSON response.

+++ internal/store/storage.go
@@ -8,7 +8,7 @@ import (
 type Storage struct {
 	Posts interface {
 		Create(context.Context, *Post) error
-		GetPostById(context.Context, int) (*Post, error)
+		GetPostById(context.Context, int64) (*Post, error)
 	}
 	Users interface {
 		Create(context.Context, *User) error
